+++ AmberCastle.Cbr.CbrWebServ/DailyInfoWebServClient.cs
@@ -1,16 +1,14 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Xml.Linq;
-using System.Xml.XPath;
+using AmberCastle.Cbr.CbrWebServ.Models;
 
 namespace AmberCastle.Cbr.CbrWebServ
 {
